+++ packages/react-icons-atomic-webpack-loader/test/src/icon-granularity-svg.js
@@ -1,0 +1,3 @@
+import { AddFilled, AddRegular, DrawImage24Filled } from '@fluentui/react-icons';
+
+export { AddFilled, AddRegular, DrawImage24Filled };